Match class names in class list to those in command results
