+++ src/main/java/seedu/taassist/model/Model.java
@@ -2,7 +2,6 @@ package seedu.taassist.model;
 
 import java.nio.file.Path;
 import java.util.Collection;
-import java.util.List;
 import java.util.Set;
 import java.util.function.Predicate;
