+++ src/test/java/seedu/taassist/logic/commands/AddcCommandTest.java
@@ -150,5 +150,11 @@ public class AddcCommandTest {
         public void addModuleClasses(Set<ModuleClass> moduleClasses) {
             requireAllNonNull(moduleClasses);
         }
+
+        @Override
+        public ModuleClass getModuleClassWithSameName(ModuleClass moduleClass) {
+            requireNonNull(moduleClass);
+            return moduleClass;
+        }
     }
 }
