+++ src/test/java/seedu/taassist/logic/commands/DeletecCommandTest.java
@@ -26,6 +26,7 @@ import seedu.taassist.model.ModelManager;
 import seedu.taassist.model.UserPrefs;
 import seedu.taassist.model.moduleclass.ModuleClass;
 import seedu.taassist.model.stubs.ModelStub;
+import seedu.taassist.model.stubs.ModelStubWithNoModuleClass;
 import seedu.taassist.model.student.Student;
 import seedu.taassist.model.uniquelist.UniqueList;
 import seedu.taassist.testutil.ModuleClassBuilder;
@@ -133,41 +134,6 @@ public class DeletecCommandTest {
 
     //==================================== Model Stubs ===============================================================
 
-    private static class ModelStubWithNoModuleClass extends ModelStub {
-
-        @Override
-        public void removeModuleClasses(Collection<ModuleClass> moduleClasses) {
-            requireAllNonNull(moduleClasses);
-        }
-
-        @Override
-        public void removeModuleClass(ModuleClass moduleClass) {
-            requireNonNull(moduleClass);
-        }
-
-        @Override
-        public boolean hasModuleClasses(Collection<ModuleClass> moduleClasses) {
-            requireAllNonNull(moduleClasses);
-            return false;
-        }
-
-        @Override
-        public boolean hasModuleClass(ModuleClass moduleClass) {
-            requireNonNull(moduleClass);
-            return false;
-        }
-
-        @Override
-        public ObservableList<ModuleClass> getModuleClassList() {
-            return new UniqueList<ModuleClass>().asUnmodifiableObservableList();
-        }
-
-        @Override
-        public ObservableList<Student> getStudentList() {
-            return new UniqueList<Student>().asUnmodifiableObservableList();
-        }
-    }
-
     private static class ModelStubWithFixedModuleClasses extends ModelStub {
         private Set<ModuleClass> moduleClasses = new HashSet<>(Arrays.asList(CS1101S, CS1231S));
 
@@ -190,6 +156,11 @@ public class DeletecCommandTest {
         @Override
         public ObservableList<Student> getStudentList() {
             return new UniqueList<Student>().asUnmodifiableObservableList();
+        }
+
+        @Override
+        public ModuleClass getModuleClassWithSameName(ModuleClass moduleClass) {
+            return moduleClass;
         }
     }
 
@@ -245,6 +216,11 @@ public class DeletecCommandTest {
         @Override
         public void setStudent(Student oldStudent, Student newStudent) {
             this.student = newStudent;
+        }
+
+        @Override
+        public ModuleClass getModuleClassWithSameName(ModuleClass moduleClass) {
+            return moduleClass;
         }
     }
 }
